***First commit after git rebase.***
The mapper will no longer be included as a git subtree. Instead, add an update script which downloads the mapper and keeps it up to date with GitHub releases.
The update script is written in Lua, though for the moment I have decided to use CURL (built as an executable) for downloading the mapper rather than Lua-CURL or LuaSocket with OpenSSL.
The included curl.exe was compiled by me using MinGW32 in order to reduce file size. The official releases of CURL statically link against OpenSSL and a bunch of other libraries which aren't needed for the purpose of running the update script. Currently, official releases of curl.exe are around 3.5MB in size, while my version which uses the built-in SSL support from Windows is only 800KB. The down side to this is that this version of CURL might fail to download from GitHub on Windows XP or older. If this is the case, the user should download an official release of CURL so that OpenSSL is used rather than WinSSL.
The update script also makes use of unzip.exe from the Info Zip package for extracting mapper releases. The included unzip.exe was taken from the official release at ftp://ftp.info-zip.org/pub/infozip/win32/unz600xn.exe
The licenses for CURL and Info Zip can be found in the docs directory of this repository.