Add some new utility functions (int, len, and spairs) to mystdlib.lua.
mume_portal_wizard:
Refactor the code for readability.
Change the trigger for receiving keys to more securely match the output of Locate Life.
Better argument parsing in the 'cache' alias.
Reformat the output of the Find The Path spell to be more readable and easier to understand with a screen reader. Output such as 'nnueee' is transformed into '2n, u, 3e'.
Reformat the output of the Locate Life spell to be less spammy. Mobs with the same key are grouped together on a single line, delineated by a comma and a space. Multiple mobs with the same name and key are represented by the mob name, preceded by the number of mobs with that name and the letter 'X'. The lines are then sorted by distance in descending order (nearest last) to make them easily reviewed with a screen reader.
When printing the list of temporary keys, sort the keys by distance in descending order.
When printing the list of cached keys, sort the keys by time elapsed since the key was captured.
Automatically clear cached keys older than 12 hours. I have seen keys still work after 14 hours and sometimes even 18 hours, but wouldn't trust a key that old for real use.
Automatically pass the name of the cached key to mume when casting Watch Room. This lets the player see the output from the watched room, preceded by the name of the key.
Fix miscellaneous bugs.

Update ReadMe file.